requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
email-validator==2.1.0
argon2-cffi==23.1.0
